+++ Locales/enUS.lua
@@ -46,4 +46,7 @@ L["OPTIONS_DESCRIPTION_WINDOW_WIDTH"] = "Set the window width of the enhanced me
                                         "size is %d. Make sure your displayed columns fit in the window!"
 L["OPTIONS_TITLE_COLUMNS"] = "Displayed Columns"
 L["OPTIONS_DESCRIPTION_COLUMNS"] = "Select the columns to be shown on merchants."
+L["OPTIONS_TITLE_COLUMN_OPTIONS"] = "Column Options"
+L["OPTIONS_TITLE_COLUMN_ITEM_SHOW_STACK"] = "Show stack size"
+
 L["ERROR_FALSE_COLLECTABLE_PET"] = "%s (%d) on this merchant is mislabeled as a pet. Please report this issue via %s"